JPA 사용하여 DB 연동 테스트

+++ server/src/com/riding/springboot/repository/UserRepository.java
@@ -1,7 +1,8 @@
 package com.riding.springboot.repository;
 
-import com.riding.springboot.jpa.User;
+import com.riding.springboot.domain.User;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<User, Long> {
+// 하나의 테이블이라고 생각하면 쉽다.
+public interface UserRepository extends CrudRepository<User, Integer>{
 }
